\name{pheatmap-package}
\alias{pheatmap-package}
\docType{package}
\title{
Pretty heatmaps
}
\description{
The package for drawing pretty heatmaps in R. The ordinary heatmap function in R has several drawbacks, when it comes to producing publication quality heatmaps. It is hard to produce pictures with consistent text, cell and overall sizes. The function pheatmap tries to alleviate the problems by offering more fine grained control over heatmap dimensions and appearance.
}
\details{
\tabular{ll}{
Package: \tab pheatmap\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-10-23\cr
License: \tab GPL3\cr
LazyLoad: \tab yes\cr
}
There is only one important function in the package \code{pheatmap}. This draws the heatmap and performs all clustering etc.
}
\author{
Raivo Kolde

Maintainer: Raivo kolde <rkolde@gmail.com>
}

\keyword{ hplot }

