% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{pheatmap-package}
\alias{pheatmap-package}
\title{Pretty heatmaps}
\description{
The package for drawing pretty heatmaps in R. The ordinary heatmap function in R has
several drawbacks when it comes to producing publication quality heatmaps. It is hard
to produce pictures with consistent text, cell and overall sizes. The function
pheatmap tries to alleviate the problems by offering more fine grained control over
heatmap dimensions and appearance.
}

